Splines.BSpline — Type

BSpline(degree, knots, ctrlpts)

Construct a b-spline object

Arguments

deg::Integer: degree

knots::Vector{Float64}:: a knot vector (u0, ... un+1)

ctrlpts::Vector{Vector{Float64}}:: control points. outer index is number of control points, inner index of dimensionality of point.

Splines.NURBS — Type

NURBS(degree, knots, weights, ctrlpts)

Construct a NURBS object

Arguments

deg::Integer: degree

knots::Vector{Float64}:: a knot vector (u0, ... un+1)

weights::Vector{Float64}:: a corresponding vector of weights

ctrlpts::Vector{Vector{Float64}}:: control points. outer index is number of control points, inner index of dimensionality of point.

Splines.curvederivativecontrolpoints — Method

curvederivativecontrolpoints(n, p, U, P, d, r1, r2)

Compute control points of curve derivatives:

\[\mathbf{C}^{(k)}(u) = \sum_{i=0}^{n-k}N_{i,p-k}(u) \mathbf{P}_i^{(k)}\]

with

\[\mathbf{P}_i^{(k)} = \begin{cases} \mathbf{P}_i & k=0 \\ \frac{p-k+1}{u_{i+p+1}-u_{i+k}}\left(\mathbf{P}_{i+1}^{(k)} - \mathbf{P}_i^{(k)} \right) & k > 0 \end{cases}\]

(see NURBS, eqn 3.8 and A3.3)

Arguments

bspline::BSpline: bspline object

`r1::Integer : first control point index to take derivatives at

`r2::Integer : last control point index to take derivatives at

d::Float: derivative order (0 ≤ k ≦ d)

Returns

PK::Matrix{Vector{Float}}: PK[i, j] is the ith derivative of the jth control point

Splines.curvederivatives — Method

curvederivatives(bspline, u, d)

Compute a curve point and its derivatives up do the dth derivative at parametric point, $u$. (NURBS, A3.2)

Arguments

bspline::BSpline: bspline object

u::Float: point on spline to evaluate at

d::Float: derivative order (0 ≤ k ≤ d)

Returns

CK::Vector{Vector{Float}}. where CK[0] is the point, CK[1] the first derivative, and so on.

Splines.curvederivatives — Method

curvederivatives(nurbs, u, d)

Compute a curve point and its derivatives up do the dth derivative at parametric point, $u$. (NURBS, A3.2)

Arguments

nurbs::NURBS: bspline object

u::Float: point on spline to evaluate at

d::Float: derivative order (0 ≤ k ≤ d)

Returns

CK::Vector{Vector{Float}} where CK[1] is the point, CK[2] the first derivative, and so on.

Splines.curvepoint — Method

Evaluate point on B-spline curve (NURBS, A3.1)

Arguments

bspline::BSpline: bspline object

u::Float: point on spline to evaluate at

Returns

C::Vector{Float}: point in ND space

Splines.curvepoint — Method

Evaluate point on rational b-spline curve (NURBS, A4.1)

Arguments

nurbs::NURBS: NURBS object

u::Float: point on spline to evaluate at

Returns

C::Vector{Float}: point in ND space

Splines.getspanindex — Method

getspanindex(deg, knots, u)

(private function) binary search to find span index of vector, knots, in which the parametric point, u, lies. (NURBS A2.1)

Arguments

deg::Integer: degree

knots::Vector{Float64}:: a knot vector (u0, ... un+1)

u::Float64`: nondimensional location we are searching for

Returns

span::Integer: corresponding index i for u between knotsi and knotsi+1

Splines.globalcurveinterpolation — Method

globalcurveinterpolation(pts, deg)

Interpolate ctrl points pts, with a B-Spline of degree deg. (NURBS A9.1)

Arguments

pts::Vector{Vector{Float}}: outer vector of length npts, inner vector of length dimension of space

deg::Integer: degree of B-spline

Outputs:

bspline::BSpline: bspline object

Splines.leastsquarescurve — Method

leastsquarescurve(pts, nctrl, deg)

Least squares fit to provided points. (NURBS section 9.4.1) TODO: Currently hard-coded to 2D data.

Arguments

pts::Vector{Vector{Float}}: data points

nctrl::Integer: number of control points to use in fit

deg::Integer: degree of bspline used in fit

Returns

bspline::BSpline: a BSpline object

Splines.singlebasisfunction — Method

singlebasisfunction(i, deg, knots, u)

(private function) Compute single basis function N_i^p. (NURBS A2.4)

Arguments

i::Integer : the index of the basis (the i in N_i)

deg::Integer : the basis degree up to the the curve order

knots::Vector{Float} : the knot vector

u::Float : parametric point of interest

Returns

Nip::Float: the N_i^p basis function value